Mv pose transformation into pose_calculations + upAxis->up_axis
Solves issue #2
Also unifies the member fields to one consisten style snake vs camel
